+++ scripts/tests/-Restart pending.ps1
@@ -1,4 +1,4 @@
-#v1.08
+#v1.09
 #https://github.com/texhex/testutil/wiki/_fwLinkScript
 
 
@@ -9,7 +9,7 @@
 Set-StrictMode -version 2.0
 
 #Terminate script on errors 
-$ErrorActionPreference = 'Stop'$Result = @{Data = "OK"; Name="No restart pending"; Text= "No pending changes that require a restart have been detected"} #Original Code [Determine Pending Reboot Status—PowerShell Style! Part 2](http://blogs.technet.com/b/heyscriptingguy/archive/2013/06/11/determine-pending-reboot-status-powershell-style-part-2.aspx) #by [Brian Wilhite](https://social.technet.microsoft.com/profile/brian%20wilhite/) #Note: PowerShell + Registry = DRAIN BAMAGED! #Get-ChildItem -> Returns the PATHS below the given path #Get-ItemProperty -> Returns the VALUES below the given path  #Check [CBS](http://technet.microsoft.com/en-us/library/cc756291%28v=ws.10%29.aspx) for Windows service packs etc. #This is a single value $cbsrebootpending=$false $test=Get-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Component Based Servicing" -Name "RebootPending" -ErrorAction Ignore  if ($test -ne $null) {
+$ErrorActionPreference = 'Stop'$Result = @{Data = "OK"; Name="No changes pending"; Text= "No pending changes that require a restart have been detected"} #Original Code [Determine Pending Reboot Status—PowerShell Style! Part 2](http://blogs.technet.com/b/heyscriptingguy/archive/2013/06/11/determine-pending-reboot-status-powershell-style-part-2.aspx) #by [Brian Wilhite](https://social.technet.microsoft.com/profile/brian%20wilhite/) #Note: PowerShell + Registry = DRAIN BAMAGED! #Get-ChildItem -> Returns the PATHS below the given path #Get-ItemProperty -> Returns the VALUES below the given path  #Check [CBS](http://technet.microsoft.com/en-us/library/cc756291%28v=ws.10%29.aspx) for Windows service packs etc. #This is a single value $cbsrebootpending=$false $test=Get-ItemProperty -Path "HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Component Based Servicing" -Name "RebootPending" -ErrorAction Ignore  if ($test -ne $null) {
      write-output "Component Based Servicing has value RebootPending set"
      $cbsrebootpending = $true
  }